+++ src/main/java/com/idrust/bmfpriceapi/services/QuandlService.java
@@ -2,15 +2,17 @@ package com.idrust.bmfpriceapi.services;
 
 import com.idrust.bmfpriceapi.exceptions.QuandlAPIException;
 
+import java.math.BigDecimal;
+
 public interface QuandlService {
 
     /**
-     * Calcula qual o preço em USD da cultura na data informada
+     * Calcula qual o preco em USD da cultura na data informada
      *
-     * @param cropCode  Código da cultura que será procurada
-     * @param date      Data em que se deseja o preço daquela cultura
-     * @return O preço em dólares (USD) da cultura com o código informado na data informada
+     * @param cropCode  Codigo da cultura que será procurada
+     * @param date      Data em que se deseja o preco daquela cultura
+     * @return O preco em dolares (USD) da cultura com o codigo informado na data informada
      */
-    Double getCropPrice(String cropCode, String date) throws QuandlAPIException;
+    BigDecimal getCropPrice(String cropCode, String date) throws QuandlAPIException;
 
 }
